feat(elec-approval): 문서 상태 검증 로직 구현 및 메소드로 분리 - 승인/반려/취소 시 문서 상태 유효성 검증 추가하여 예외 처리 - validateDocumentIsPendingOrInProgress 메소드로 분리하여 재사용

+++ coretime/src/main/java/com/flow/coretime/elecApproval/controller/ElecApprovalController.java
@@ -260,7 +260,7 @@ public class ElecApprovalController {
         @PreAuthorize("@documentChecker.isInitiator(#p0, principal.username)")
         @PostMapping("/recall/{docId}")
         @ResponseBody
-        public ResponseEntity<ApiResponse<Void>> recall(@PathVariable(name = "docId") Long docId,
+        public ResponseEntity<ApiResponse<Void>> recall(@PathVariable(name = "docId") int docId,
                         @AuthenticationPrincipal UserDetails userDetails) {
                 elecApprovalCommandService.recallDocument(docId, userDetails.getUsername());
                 return ResponseEntity.ok(ApiResponse.success("상신 취소가 성공적으로 완료되었습니다."));
